feature(difficulty-level): change garbage manager next position
